+++ packages/ui/src/index.js
@@ -7,17 +7,17 @@ import * as mixins from './mixins';
 import resetAntStyles from './antd';
 
 // Data
-import * as footerData from './components/footer/footer.data';
-import * as headerData from './components/header/header.data';
+import footerData, * as defaultFooterData from './components/footer/footer.data';
+import headerData, * as defaultHeaderData from './components/header/header.data';
 // Utils
 import * as headerUtils from './components/header/header.utils';
 import { nav } from './utils';
 
-export const getData = (domain, path) => {
+export const getData = (domain, path, locale) => {
   const urlPrefixRegexp = nav.buildUrlPrefixPattern(domain, path);
 
-  const { navItems: footerNavItems, ...restFooterData } = footerData;
-  const { navItems: headerNavItems, ...restHeaderData } = headerData;
+  const { navItems: footerNavItems, ...restFooterData } = footerData[locale] || defaultFooterData;
+  const { navItems: headerNavItems, ...restHeaderData } = headerData[locale] || defaultHeaderData;
 
   return {
     footer: {