+++ src/layouts/index.js
@@ -1,2 +1,1 @@
 export { default as CoreLayout } from './core';
-export { default as LegacyLayout } from './legacy';