+++ packages/ui/CHANGELOG.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## [0.1.1] - 2021-04-15
+
+### Break changes
+
+- Add `hasMargin` prop controlling footer margin style
+- Use `getData` to remove url prefix for current site in nav links
+
+### Added
+
+- Add temporary solution for i18n in header
+
+---
+
 ## [0.1.0] - 2021-04-12
 
 ### Added